requirements.txt
google-cloud-bigquery>=3.11.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0